+++ src/network/utils/mac16-address.cc
@@ -106,6 +106,7 @@ Mac16Address::Mac16Address(const char* str)
 
 Mac16Address::Mac16Address(uint16_t addr)
 {
+    NS_LOG_FUNCTION(this);
     m_address[1] = addr & 0xFF;
     m_address[0] = (addr >> 8) & 0xFF;
 }
@@ -151,6 +152,15 @@ Mac16Address::ConvertTo() const
 {
     NS_LOG_FUNCTION(this);
     return Address(GetType(), m_address, 2);
+}
+
+uint16_t
+Mac16Address::ConvertToInt() const
+{
+    uint16_t addr = m_address[1] & (0xFF);
+    addr |= (m_address[0] << 8) & (0xFF << 8);
+
+    return addr;
 }
 
 Mac16Address